Update and rename README.md to convolution_imageprocessing_code.m
This MATLAB code performs convolution on an image using a predefined kernel. The image is read using the imread function and displayed using the imshow function. The kernel is defined as a 3x3 matrix and stored in an array named SE.

The code then initializes a new array named new to store the result of the convolution operation. A nested for loop is used to iterate through the pixels of the input image, and for each pixel, a 3x3 sub-matrix is extracted from the input image. This sub-matrix is then multiplied element-wise with the kernel matrix SE, and the sum of the resulting matrix is computed. This sum is then stored in the corresponding position in the output matrix new.

The resulting new matrix is then converted to an 8-bit unsigned integer using the uint8 function so that it can be displayed using the imshow function. The output image is displayed in a separate plot with the title "after convolution". Overall, this code demonstrates how to perform convolution on an image using a custom kernel in MATLAB.